Fix unit test compilation under HIP

+++ test/lbvh_test.cu
@@ -134,8 +134,8 @@ UP_TEST( lbvh_test )
             int4 node = data[0];
             UP_ASSERT_EQUAL(((unsigned int)node.x >> 20) & 0x3ffu, 0);
             UP_ASSERT_EQUAL(((unsigned int)node.y >> 20) & 0x3ffu, 0);
-            UP_ASSERT_EQUAL(node.z, 1);
-            UP_ASSERT_EQUAL(node.w, neighbor::LBVHSentinel);
+            UP_ASSERT_EQUAL((int)node.z, 1);
+            UP_ASSERT_EQUAL((int)node.w, neighbor::LBVHSentinel);
             }
         // Node 1
             {
@@ -143,8 +143,8 @@ UP_TEST( lbvh_test )
 
             UP_ASSERT_EQUAL(((unsigned int)node.x >> 20) & 0x3ffu, 0);
             UP_ASSERT_EQUAL(((unsigned int)node.y >> 20) & 0x3ffu, 511);
-            UP_ASSERT_EQUAL(node.z, 2);
-            UP_ASSERT_EQUAL(node.w, 4);
+            UP_ASSERT_EQUAL((int)node.z, 2);
+            UP_ASSERT_EQUAL((int)node.w, 4);
             }
         // Node 2
             {
@@ -152,8 +152,8 @@ UP_TEST( lbvh_test )
 
             UP_ASSERT_EQUAL(((unsigned int)node.x >> 20) & 0x3ffu, 0);
             UP_ASSERT_EQUAL(((unsigned int)node.y >> 20) & 0x3ffu, 1023);
-            UP_ASSERT_EQUAL(node.z, ~2);
-            UP_ASSERT_EQUAL(node.w, 3);
+            UP_ASSERT_EQUAL((int)node.z, ~2);
+            UP_ASSERT_EQUAL((int)node.w, 3);
             }
         // Node 3
             {
@@ -161,8 +161,8 @@ UP_TEST( lbvh_test )
 
             UP_ASSERT_EQUAL(((unsigned int)node.x >> 20) & 0x3ffu, 511);
             UP_ASSERT_EQUAL(((unsigned int)node.y >> 20) & 0x3ffu, 511);
-            UP_ASSERT_EQUAL(node.z, ~1);
-            UP_ASSERT_EQUAL(node.w, 4);
+            UP_ASSERT_EQUAL((int)node.z, ~1);
+            UP_ASSERT_EQUAL((int)node.w, 4);
             }
         // Node 4
             {
@@ -170,8 +170,8 @@ UP_TEST( lbvh_test )
 
             UP_ASSERT_EQUAL(((unsigned int)node.x >> 20) & 0x3ffu, 1023);
             UP_ASSERT_EQUAL(((unsigned int)node.y >> 20) & 0x3ffu, 0);
-            UP_ASSERT_EQUAL(node.z, ~0);
-            UP_ASSERT_EQUAL(node.w, neighbor::LBVHSentinel);
+            UP_ASSERT_EQUAL((int)node.z, ~0);
+            UP_ASSERT_EQUAL((int)node.w, neighbor::LBVHSentinel);
             }
 
         // each node should have the correct number of hits
@@ -556,8 +556,8 @@ UP_TEST( lbvh_small_test )
         // only one node, just check its index contents
             {
             int4 node = data[0];
-            UP_ASSERT_EQUAL(node.z, ~0);
-            UP_ASSERT_EQUAL(node.w, neighbor::LBVHSentinel);
+            UP_ASSERT_EQUAL((int)node.z, ~0);
+            UP_ASSERT_EQUAL((int)node.w, neighbor::LBVHSentinel);
             }
 
         // each node should have the correct number of hits
